Add User Type Allowed
I did this by creating a new table called parking restrictions, it maps
lot_id to type of user allowed. Its a Many to one relationship, so it
is expected that there will be multiple entries for a singel parking lot,
if the lot allows different types of users. My reason for doing it like this
and not simply adding an attribute to the parking_lot_information table
is to keep the database in normalized.